Fix division by 0.0 in BaselineRow::PerpDistanceFromBaseline
It was reported by oss-fuzz (issue 23962).

Add log output to find real images which trigger that issue.
Avoid also some conversions from float to double by always using float.

+++ src/textord/baselinedetect.h
@@ -95,7 +95,7 @@ class BaselineRow {
                               double target_offset);
   // Returns the perpendicular distance of the point from the straight
   // baseline.
-  double PerpDistanceFromBaseline(const FCOORD& pt) const;
+  float PerpDistanceFromBaseline(const FCOORD& pt) const;
   // Computes the bounding box of the row.
   void ComputeBoundingBox();
